requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow>=9.1.0,<9.2
django-user-agents>=0.4.0,<0.4.1
django-cors-headers>=4.3.1,<4.4
requests>=2.26.0,<2.27
